rename(useCardNumber): 함수 위치 분리
